[#32] refactor: 카테고리 왼쪽 사이드 바 컴포넌트 명 및 전체적인 구조 변경
